requirements.txt
aiogram==3.26.0
aiosqlite>=0.20,<1.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
